web-api/requirements.txt
click==8.1.2
marshmallow>=3
factory_boy
flask==2.0.2
flask-httpauth==4.5.0
flask-restful==0.3.9
flask-migrate==3.1.0
# flask-redis==0.4.0
flask-sqlalchemy==2.5.1
flask-swagger-ui==3.36.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0
passlib==1.7.4
psycopg2==2.9.3
python-dotenv==0.19.2
requests==2.27.1
python-decouple==3.5
apispec==5.1.1
apispec-webframeworks==0.5.2
# redis==4.1.0
pyjwt==2.3.0
gunicorn==20.0.4
pyyaml==6.0
# mockredispy==2.9.3
